Fix: seleção correta do template - template details
