+++ src/components/TemplateDetails/Summary.vue
@@ -10,18 +10,14 @@
           Mensagens enviadas
           <unnnic-icon icon="info" size="sm" filled scheme="neutral-cleanest" />
         </div>
-        <div class="whatsapp_template_summary__main__card__value">
-          {{ templateAnalytics.data[0].totals.sent || 0 }}
-        </div>
+        <div class="whatsapp_template_summary__main__card__value">{{ weekValues.sent }}</div>
       </div>
       <div class="whatsapp_template_summary__main__card">
         <div class="whatsapp_template_summary__main__card__title">
           Mensagens entregues
           <unnnic-icon icon="info" size="sm" filled scheme="neutral-cleanest" />
         </div>
-        <div class="whatsapp_template_summary__main__card__value">
-          {{ templateAnalytics.data[0].totals.delivered || 0 }}
-        </div>
+        <div class="whatsapp_template_summary__main__card__value">{{ weekValues.delivered }}</div>
       </div>
       <div class="whatsapp_template_summary__main__card">
         <div class="whatsapp_template_summary__main__card__title">
@@ -29,19 +25,16 @@
           <unnnic-icon icon="info" size="sm" filled scheme="neutral-cleanest" />
         </div>
         <div class="whatsapp_template_summary__main__card__value">
-          {{ templateAnalytics.data[0].totals.read || 0 }} ({{
-            Math.floor(
-              (templateAnalytics.data[0].totals.read / templateAnalytics.data[0].totals.sent) * 100,
-            ) || '--'
-          }}
-          %)
+          {{ weekValues.read }} ({{ Math.floor((weekValues.read / weekValues.sent) * 100) }}%)
         </div>
       </div>
     </div>
     <div class="whatsapp_template_summary__quality">
       <div class="whatsapp_template_summary__quality__title">
         <div class="whatsapp_template_summary__title__main">Visão geral da qualidade</div>
-        <div class="whatsapp_template_summary__title__subtitle">Últimos 7 dias</div>
+        <div class="whatsapp_template_summary__title__subtitle">
+          Últimos {{ templateAnalytics.data.lenght }} dias
+        </div>
       </div>
       <div class="whatsapp_template_summary__quality__status">
         <div class="whatsapp_template_summary__quality__status__level">
@@ -60,13 +53,43 @@
 </template>
 
 <script>
-  import { mapState } from 'vuex';
+  import { mapState, mapActions } from 'vuex';
   export default {
     name: 'Summary',
+    mounted() {
+      this.fetchTemplateAnalyticsWeek();
+    },
     computed: {
       ...mapState({
         templateAnalytics: (state) => state.insights.templateAnalytics,
+        appUuid: (state) => state.insights.appUuid,
+        selectedTemplate: (state) => state.insights.selectedTemplate,
       }),
+      weekValues() {
+        return this.templateAnalytics.grand_totals;
+      },
+    },
+    methods: {
+      ...mapActions(['getTemplateAnalytics', 'getTemplates']),
+      fetchTemplateAnalyticsWeek() {
+        const params = {
+          app_uuid: this.appUuid,
+          filters: {
+            start: this.formatDate(new Date(Date.now() - 7 * 24 * 60 * 60 * 1000)),
+            end: this.formatDate(new Date()),
+            fba_template_ids: this.selectedTemplate.translations.map(
+              (item) => item.message_template_id,
+            ),
+          },
+        };
+        this.getTemplateAnalytics(params);
+        if (this.errorTemplateAnalytics) {
+          alert(this.errorTemplateAnalytics);
+        }
+      },
+      formatDate(date) {
+        return `${date.getMonth() + 1}-${date.getDate()}-${date.getFullYear()}`;
+      },
     },
   };
 </script>
